+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added array support to the `with_byline_manager_authors()` and `with_cap_authors()` methods on post factories.
+- Added `assertJsonPathEmpty()`, `assertJsonPathNotEmpty()`,
+  `assertJsonPathContains()`, `assertJsonPathNotContains()`, and
+  `assertJsonPathCallback()` methods to the test response class.
+
 ### Changed
 
 - Changed the `theme()` method of the installation manager to set the
@@ -15,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Throw an exception if the user attempts to make a testing request to a path
   that is not supported by the testing framework (e.g. `/wp-login.php`,
   `/wp-admin/*`, etc.).
+
+### Fixed
+
+- Fixed issue with empty JSON not being parsed properly with `Assertable_Json_String`.
 
 ## v1.9.2